@@ -11,9 +11,9 @@
   },
   "dependencies": {
     "nanoid": "^5.0.7",
-    "react": "^17.0.0",
+    "react": "^18.3.1",
     "react-confetti": "^6.1.0",
-    "react-dom": "^17.0.0",
+    "react-dom": "^18.3.1",
     "react-use": "^17.5.1"
   },
   "devDependencies": {
