@@ -1,11 +1,13 @@
-import { StrictMode } from 'react'
-import ReactDOM from 'react-dom'
 import App from './App.jsx'
+import { createRoot } from 'react-dom/client'
+import { StrictMode } from 'react'
 import './index.css'
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <StrictMode>
     <App />
-  </StrictMode>,
-  document.getElementById('root')
-)
+  </StrictMode>
+);
